==> birds_data/2 Get info from oiseaux-birds.ipynb <==
# import pandas
import pandas as pd
# import urllib to deal with url connection
from urllib.request import Request, urlopen
# import beautiful soup to extract information from a webpage
from bs4 import BeautifulSoup as bs
# import regular expression lib
import re
# import multiprocessing lib to accelerate fetching data, windows does not support this lib
from multiprocessing import Pool
import string

# read birds list into a dataframe
birds_list = pd.read_csv("birds_list.csv")

# check the head of the birds list
birds_list.head()

# convert the format of common name into lowercase with underscore, like 'australian_brushturkey'
birds_list['Common_name_lowercase'] = birds_list['Common name'].str.lower()
birds_list['Common_name_lowercase'] = birds_list['Common_name_lowercase'].str.replace(' ', '-')

# gerenate all the links to http://www.oiseaux-birds.com/ for every bird
link_prefix = 'http://www.oiseaux-birds.com/'
birds_list['link'] = link_prefix + 'card-' + birds_list['Common_name_lowercase'] + '.html'

# if one bird cannot found in http://www.oiseaux-birds.com/, then there will be a error page which length is 2282.
ERROR_PAGE_LEN = 2282

# try to get information for one specific bird

url = "http://www.oiseaux-birds.com/card-antipodean-albatross.html"
# use Firefox as an user-agent, otherwise the connection will be rejected.
req = Request(url, headers={'User-Agent': 'Mozilla/5.0'})
# get the content of a webpage
webpage = urlopen(req).read()
# convert to string
webpage_str = str(webpage)
# initialize a beautiful soup with lxml engine
soup = bs(webpage, "lxml")

# use regular expression to get height, it it doesn't exist, leave it blank.
height = re.search('Height:(.*?)<br />', webpage_str)
if height:
    height = height.group(1).strip()
else:
    height = pd.np.nan

# use regular expression to get length, it it doesn't exist, leave it blank.
length = re.search('Length:(.*?)<br />', webpage_str)
if length:
    length = length.group(1).strip()
else:
    length = pd.np.nan

# use regular expression to get weight, it it doesn't exist, leave it blank.
weight = re.search('Weight:(.*?)</p>', webpage_str)
if weight:
    weight = weight.group(1).strip()
else:
    weight = pd.np.nan

# find all img tags in html file
imgs = soup.find_all("img")
if imgs:
    # list comprehension to have a list of complete link of img files
    imgs = [link_prefix + img['src'] for img in imgs]
else:
    imgs = pd.np.nan

# find the position of voice link by regular expression
voice = re.search('VOICE(.*?)</a>', webpage_str)
if voice:
    voice = voice.group(1).strip()
    voice = re.search('<a href="(.*?)"', voice)
    if voice:
        voice = voice.group(1).strip()
        
        # The parent of parent of parent of a tag contains 'VOICE' is where the voice link is
        for elem in soup(text=re.compile(r'VOICE')):
            voice_text = elem.parent.parent.parent.get_text()
            voice_text = voice_text[6:].strip()
            #print(voice_text)
            #print(voice)
else:
    voice = pd.np.nan
    voice_text = pd.np.nan

# extract habitat by beautiful soup, if can not found, leave it blank
habitat = re.search('HABITAT', webpage_str)
if habitat:
    for elem in soup(text=re.compile(r'HABITAT')):
        habitat = elem.parent.parent.parent.get_text()
        habitat = habitat[8:].strip()
        print(habitat)
else:
    habitat = pd.np.nan

# extract description by beautiful soup, if can not found, leave it blank
description = re.search('DESCRIPTION OF THE BIRD', webpage_str)
if description:
    for elem in soup(text=re.compile(r'DESCRIPTION OF THE BIRD')):
        description = elem.parent.parent.parent.parent.get_text()
        description = description[25:].strip()
        print(description)
else:
    description = re.search('DESCRIPTION', webpage_str)
    if description:
        for elem in soup(text=re.compile(r'DESCRIPTION')):
            description = elem.parent.parent.parent.get_text()
            description = description[12:].strip()
            print(description)
    else:
        description = pd.np.nan

if description == description:
    find_biometrics = re.search('Biometrics((?!to).|\n)*\d *(k?gr? *(\(.*\))?|cm *(\(.*\))?)\.?', description)
    if find_biometrics:
        description = description[:find_biometrics.span()[0]] + description[find_biometrics.span()[1]:]
        description = description.strip()
    if len(description) > 0 and description[0] not in string.ascii_letters:
        description = description[1:]
    description = description.strip()

# extract movements by beautiful soup, if can not found, leave it blank
movements = re.search('MOVEMENTS', webpage_str)
if movements:
    for elem in soup(text=re.compile(r'MOVEMENTS')):
        movements = elem.parent.parent.parent.get_text()
        movements = movements[10:].strip()
        print(movements)
else:
    movements = pd.np.nan

# extract reproduction by beautiful soup, if can not found, leave it blank
reproduction = re.search('REPRODUCTION', webpage_str)
if reproduction:
    for elem in soup(text=re.compile(r'REPRODUCTION')):
        reproduction = elem.parent.parent.parent.get_text()
        reproduction = reproduction[13:].strip()
        print(reproduction)
else:
    reproduction = pd.np.nan

# extract diet by beautiful soup, if can not found, leave it blank
diet = re.search('DIET', webpage_str)
if diet:
    for elem in soup(text=re.compile(r'DIET')):
        diet = elem.parent.parent.parent.get_text()
        diet = diet[5:].strip()
        print(diet)
else:
    diet = pd.np.nan

# extract protection by beautiful soup, if can not found, leave it blank
protection = re.search('PROTECTION', webpage_str)
if protection:
    for elem in soup(text=re.compile(r'PROTECTION')):
        protection = elem.parent.parent.parent.get_text()
        protection = protection[31:].strip()
        print(protection)
else:
    protection = pd.np.nan

# extract sources by beautiful soup, if can not found, leave it blank
sources = re.search('Sources', webpage_str)
if sources:
    for elem in soup(text=re.compile(r'Sources')):
        sources = elem.parent.parent.parent.parent.get_text()
        sources = sources[sources.find("Sources") + 9:sources.find("Home page")].strip()
        print(sources)
else:
    sources = pd.np.nan

# define a error_list, if one bird can not find in this website, then add it into this list
error_list = []

# define a fetch function to get all the information for a multiprocessing pool
def fetch(idx):
    url = birds_list.ix[idx, 'link']
    req = Request(url, headers={'User-Agent': 'Mozilla/5.0'})
    webpage = urlopen(req).read()
    webpage_str = str(webpage)
    soup = bs(webpage, "lxml")
    imgs_dict = dict()
    
    if len(webpage) == ERROR_PAGE_LEN:
        error_list.append(idx)
        return pd.np.nan
    
    height = re.search('Height:(.*?)<br />', webpage_str)
    if height:
        height = height.group(1).strip()
    else:
        height = pd.np.nan
    
    length = re.search('Length:(.*?)<br />', webpage_str)
    if length:
        length = length.group(1).strip()
    else:
        length = pd.np.nan

    weight = re.search('Weight:(.*?)</p>', webpage_str)
    if weight:
        weight = weight.group(1).strip()
    else:
        weight = pd.np.nan
    
    imgs = soup.find_all("img")
    if imgs:
        imgs = [link_prefix + img['src'] for img in imgs]
    else:
        imgs = pd.np.nan
    
    voice = re.search('VOICE(.*?)</a>', webpage_str)
    if voice:
        voice = voice.group(1).strip()
        voice = re.search('<a href="(.*?)"', voice)
        if voice:
            voice = voice.group(1).strip()

            for elem in soup(text=re.compile(r'VOICE')):
                voice_text = elem.parent.parent.parent.get_text()
                voice_text = voice_text[6:].strip()
                #print(voice_text)
                #print(voice)
    else:
        voice = pd.np.nan
        voice_text = pd.np.nan
    
    habitat = re.search('HABITAT', webpage_str)
    if habitat:
        for elem in soup(text=re.compile(r'HABITAT')):
            habitat = elem.parent.parent.parent.get_text()
            habitat = habitat[8:].strip()
            #print(habitat)
    else:
        habitat = pd.np.nan
    
    description = re.search('DESCRIPTION OF THE BIRD', webpage_str)
    if description:
        for elem in soup(text=re.compile(r'DESCRIPTION OF THE BIRD')):
            description = elem.parent.parent.parent.parent.get_text()
            description = description[25:].strip()
            # print(description)
    else:
        description = re.search('DESCRIPTION', webpage_str)
        if description:
            for elem in soup(text=re.compile(r'DESCRIPTION')):
                description = elem.parent.parent.parent.get_text()
                description = description[12:].strip()
                # print(description)
        else:
            description = pd.np.nan
    if description == description:
        find_biometrics = re.search('Biometrics((?!to).|\n)*\d *(k?gr? *(\(.*\))?|cm *(\(.*\))?)\.?', description)
        if find_biometrics:
            description = description[:find_biometrics.span()[0]] + description[find_biometrics.span()[1]:]
            description = description.strip()
        if len(description) > 0 and description[0] not in string.ascii_letters:
            description = description[1:]
        description = description.strip()
    
    movements = re.search('MOVEMENTS', webpage_str)
    if movements:
        for elem in soup(text=re.compile(r'MOVEMENTS')):
            movements = elem.parent.parent.parent.get_text()
            movements = movements[10:].strip()
            #print(movements)
    else:
        movements = pd.np.nan
    
    reproduction = re.search('REPRODUCTION', webpage_str)
    if reproduction:
        for elem in soup(text=re.compile(r'REPRODUCTION')):
            reproduction = elem.parent.parent.parent.get_text()
            reproduction = reproduction[13:].strip()
            #print(reproduction)
    else:
        reproduction = pd.np.nan
    
    diet = re.search('DIET', webpage_str)
    if diet:
        for elem in soup(text=re.compile(r'DIET')):
            diet = elem.parent.parent.parent.get_text()
            diet = diet[5:].strip()
            #print(diet)
    else:
        diet = pd.np.nan
    
    protection = re.search('PROTECTION', webpage_str)
    if protection:
        for elem in soup(text=re.compile(r'PROTECTION')):
            protection = elem.parent.parent.parent.get_text()
            protection = protection[31:].strip()
            #print(protection)
    else:
        protection = pd.np.nan
    
    sources = re.search('Sources', webpage_str)
    if sources:
        for elem in soup(text=re.compile(r'Sources')):
            sources = elem.parent.parent.parent.parent.get_text()
            sources = sources[sources.find("Sources") + 9:sources.find("Home page")].strip()
            #print(sources)
    else:
        sources = pd.np.nan
    
    birds_list.ix[idx, 'height'] = height
    birds_list.ix[idx, 'length'] = length
    birds_list.ix[idx, 'weight'] = weight
    if imgs == imgs:
        imgs_dict[idx] = imgs
    birds_list.ix[idx, 'voice'] = voice
    birds_list.ix[idx, 'habitat'] = habitat
    birds_list.ix[idx, 'description'] = description
    birds_list.ix[idx, 'movements'] = movements
    birds_list.ix[idx, 'reproduction'] = reproduction
    birds_list.ix[idx, 'diet'] = diet
    birds_list.ix[idx, 'protection'] = protection
    birds_list.ix[idx, 'sources'] = sources

    imgs_dict[idx] = imgs
    
    return (birds_list.ix[idx], imgs_dict)

#### Even with 20 process at one time, checking about 900 birds still takes very long time, seperate them into 5 parts, otherwise it may crash or get timeout when executing these tasks.

# Create a new Pool object
multi_pool = Pool(processes=20)
# Extract all info
new_bird_list = multi_pool.map(fetch, range(200))
# Close Multiprocessing Pool
multi_pool.close()

# Create a new Pool object
multi_pool = Pool(processes=20)
# Extract all info
new_bird_list1 = multi_pool.map(fetch, range(200, 400))
# Close Multiprocessing Pool
multi_pool.close()

# Create a new Pool object
multi_pool = Pool(processes=20)
# Extract all info
new_bird_list2 = multi_pool.map(fetch, range(400, 600))
# Close Multiprocessing Pool
multi_pool.close()

# Create a new Pool object
multi_pool = Pool(processes=20)
# Extract all info
new_bird_list3 = multi_pool.map(fetch, range(600, 800))
# Close Multiprocessing Pool
multi_pool.close()

# Create a new Pool object
multi_pool = Pool(processes=20)
# Extract all info
new_bird_list4 = multi_pool.map(fetch, range(800, len(birds_list)))
# Close Multiprocessing Pool
multi_pool.close()

# combine all 5 parts data into one list
birds_list_cleaned = new_bird_list + new_bird_list1 + new_bird_list2 + new_bird_list3 + new_bird_list4
birds_list_cleaned = [bird for bird in birds_list_cleaned if type(bird) is not float]

# check how many birds have got in oiseaux-birds
len(birds_list_cleaned)

imgs_list = [bird[1] for bird in birds_list_cleaned]
birds_list_cleaned = [bird[0] for bird in birds_list_cleaned]

# create a new dataframe with data from oiseaux-birds
birds_new_df = pd.DataFrame(birds_list_cleaned)
birds_new_df = birds_new_df.rename(columns={'Unnamed: 0': 'id'})
birds_new_df.set_index('id', inplace=True)
birds_new_df.columns

# create a img dataframe to store all img links for each birds
imgs_dict = dict()
for imgs in imgs_list:
    imgs_dict.update(imgs)

imgs_df = pd.DataFrame(dict([(k, pd.Series(v)) for k,v in imgs_dict.items()]))

# save new dataframe to csv file
birds_new_df.to_csv("birds.csv")
imgs_df.to_csv("imgs.csv")

# create new dataframe to store the birds cannot find in oiseaux and save it to a csv file
incomplete_df = birds_list[~birds_list.index.isin(birds_new_df.index)]
incomplete_df.to_csv("can_not_find_in_oiseaux.csv")
