==> birds_data/3 Rename_column_names.ipynb <==
import pandas as pd

df = pd.read_csv("birds.csv")

del df['Common_name_lowercase']
del df['link']

df.rename(columns={'Common name': 'common_name', 'Binomial': 'binomial', 'Category': 'category', 'Order': 'order',
                   'Family': 'family'}, inplace=True)

df.columns

df.set_index('id', inplace=True)

df.head()

df.to_csv("birds.csv")
