==> birds_data/4 Combine_Colours_location_length_weight.ipynb <==
# import pandas and regular expression library
import pandas as pd
import re

# load data into dataframe
new_df = pd.read_csv("birds_with_manual_info.csv")
ori_df = pd.read_csv("birds.csv")

# set index for two DataFrame
new_df.set_index('id', inplace=True)
ori_df.set_index('id', inplace=True)

del ori_df['height']
del ori_df['length']
del ori_df['weight']

ori_df['colour1'] = new_df['colour1']
ori_df['colour2'] = new_df['colour2']
ori_df['colour3'] = new_df['colour3']

ori_df['min_length'] = new_df['min_length']
ori_df['max_length'] = new_df['max_length']
ori_df['min_weight'] = new_df['min_weight']
ori_df['max_weight'] = new_df['max_weight']

ori_df['vic'] = new_df['vic']
ori_df['nsw'] = new_df['nsw']
ori_df['qld'] = new_df['qld']
ori_df['act'] = new_df['act']
ori_df['nt'] = new_df['nt']
ori_df['sa'] = new_df['sa']
ori_df['tas'] = new_df['tas']
ori_df['wa'] = new_df['wa']

ori_df.columns

new_df.columns

# save dataframe to csv file
new_df.to_csv("birds.csv")
