==> birds_data/5 Imgs_downloader.ipynb <==
# import pandas, urllib, os library
import pandas as pd
import urllib.request
import os
from PIL import Image
import io

# read the imgs link csv
imgs_df = pd.read_csv("imgs.csv")

# remove unused column
del imgs_df['Unnamed: 0']

# get the transpose of the dataframe
imgs_T_df = imgs_df.T

# get the column names into a list
columns_name = imgs_T_df.columns

# initialize 5 set data structure, split downloading process to 5 part for the downloading processing is really slow
# and there will encounter timeout error or file cannot find
imgs_list = list(imgs_T_df.index)

directory = 'images'

def convert_imgs(img, filename):
    image = Image.open(img)
    ram = io.BytesIO()
    image.save(ram, format='png', optimize=True)
    ram.seek(0)
    im = Image.open(ram)
    im2 = im.convert('RGB').convert('P', palette=Image.ADAPTIVE)
    im2.save(directory + '/' + filename, format='png', optimize=True)

# download files
def download_imgs(imgs_list):
    for idx in imgs_list:
        if not os.path.exists(directory):
            os.makedirs(directory)
        for column in columns_name:
            img_url = imgs_T_df.ix[idx][column]
            if type(img_url) is not float:
                filename = str(idx) + "_" + str(column) + ".png"
                name_with_path = directory + '/' + filename
                if os.path.isfile(name_with_path):
                    continue
                try:
                    request = urllib.request.urlopen(img_url, timeout=5)
                    with open(name_with_path, 'wb') as f:
                        f.write(request.read())
                    convert_imgs(name_with_path, filename)
                except:
                    pass
            else:
                break

download_imgs(imgs_list[0:50])

download_imgs(imgs_list[50:100])

download_imgs(imgs_list[100:150])

download_imgs(imgs_list[150:200])

download_imgs(imgs_list[200:])
