==> birds_data/6 Get_voice.ipynb <==
# Import necessary libraries
import pandas as pd
from urllib.request import Request, urlopen
import re
import os

# load latest csv file into dataframe
birds_df = pd.read_csv("birds.csv")
birds_df.set_index('id', inplace=True)

# load latest csv file into dataframe
new_df = pd.read_csv("birds_with_more_voice.csv")
new_df.set_index('Idx', inplace=True)

# merge voice url
voice_isnull_idx = birds_df[birds_df['voice'].isnull()].index
birds_df.ix[voice_isnull_idx, 'voice'] = new_df.ix[voice_isnull_idx, 'URL to list of all the species bird calls']

# get the index of birds which has voice link
voice_idx = birds_df[birds_df['voice'].notnull()].index

# create a string containing the prefix of url to generate a complete url later.
url_prefix = "http://www.xeno-canto.org"

# download all the voice file into voice/
for idx in voice_idx:
    # as this block of code easily fail due to lose connection to the server, 
    # when restart, check it the file already existed
    if os.path.isfile("voice/" + str(idx) + '.mp3'):
        continue
    
    url = birds_df.ix[idx, 'voice']
    req = Request(url, headers={'User-Agent': 'Mozilla/5.0'})
    webpage = urlopen(req).read()
    webpage_str = str(webpage)

    # use regular expression to find where download link is
    url_surfix = re.search('/\d*/download', webpage_str)
    # if there does not exists, continue to next iteration
    if url_surfix is None:
        continue
    url_surfix = url_surfix[0]

    mp3_url = url_prefix + url_surfix
    
    # create a download command using aria2c as download tool and rename it into "[idx].mp3"
    download_command = 'aria2c ' + mp3_url + ' --out="voice/' + str(idx) + '.mp3"'
    !$download_command

birds_df.to_csv('birds.csv')
