==> birds_data/9 Fetching birds info from Wikipedia.ipynb <==
# import pandas
import pandas as pd

# import urllib to get data from webpage
from urllib.request import Request, urlopen
# import beautifulsoup, regular expression to extract data
from bs4 import BeautifulSoup as bs
import re

# Read birds list
birds_list = pd.read_csv("birds_list.csv")
birds_list = birds_list.rename(columns={'Unnamed: 0': 'Idx'})
birds_list.set_index('Idx', inplace=True)

# Read the current cleaned birds data to get the idx
birds_df = pd.read_csv("birds_April_6.csv")
birds_df.set_index('Idx', inplace=True)

# Get the missing data list
index_diff = birds_list.index.difference(birds_df.index)
birds_list = birds_list.ix[index_diff]

birds_list['common_name_in_url'] = birds_list['Common name'].map(lambda x:"_".join(x.split()))

url_prefix = "https://en.wikipedia.org/wiki/"

birds_list['url'] = url_prefix + birds_list['common_name_in_url']

index_list = list(birds_list.index)

index_list1 = index_list[:100]
index_list2 = index_list[100:200]
index_list3 = index_list[200:300]
index_list4 = index_list[300:400]
index_list5 = index_list[400:500]
index_list6 = index_list[500:600]
index_list7 = index_list[600:]

def test_url(index_sub_list):
    for idx in index_sub_list:
        url = birds_list.ix[idx, 'url']
        req = Request(url, headers={'User-Agent': 'Mozilla/5.0'})
        try:
            webpage = urlopen(req).read()
        except:
            birds_list.ix[idx, 'url'] = pd.np.nan

test_url(index_list1)

test_url(index_list2)

test_url(index_list3)

test_url(index_list4)

test_url(index_list5)

test_url(index_list6)

test_url(index_list7)

birds_list.to_csv("birds_info_wiki.csv")

birds_list.ix[39, 'url'] = 'https://en.wikipedia.org/wiki/Mallard'

birds_list.ix[354, 'url'] = 'https://en.wikipedia.org/wiki/Baird%27s_sandpiper'

birds_list = pd.concat([birds_list, pd.DataFrame(columns=['nsw', 'vic', 'tas', 'act', 'sa', 'wa', 'nt', 'qld', 
                                                          'min_weight', 'max_weight', 'min_length', 'max_length',
                                                          'images_url', 'voice_url', 'Description', 'Sources'
                                                          'colour'])])

birds_list_nul = birds_list[birds_list['url'].isnull()]

len(birds_list)

len(birds_list_nul)

birds_list_nul

birds_list.to_csv("birds_info_wiki.csv")

birds_list_nul.to_csv('birds_not_found_in_wiki.csv')
