==> birds_data/50 Split into different tables.ipynb <==
import pandas as pd
import os

csv_file = "birds_April_13.csv"
birds_df = pd.read_csv(csv_file)

birds_df.columns

birds_loc_df = birds_df[['id', 'vic', 'nsw', 'qld', 'act', 'nt', 'sa', 'tas', 'wa']].copy()

# birds_loc_df.to_json('final_set/location.json')
birds_loc_df.set_index('id', inplace=True)
birds_loc_df.to_csv('final_set/location.csv')

birds_color_df = birds_df[['id', 'colour1', 'colour2', 'colour3']].copy()

# birds_color_df.to_json('final_set/colour.json')
birds_color_df.set_index('id', inplace=True)
birds_color_df.to_csv('final_set/colour.csv')

name_list = os.listdir('voice/')[1:]
voice_id_list = [int(name.split('.')[0]) for name in name_list]

birds_voice_df = birds_df[['id', 'voice']].copy()

birds_voice_df['local'] = pd.np.nan
birds_voice_df['voice'] = pd.np.nan
birds_voice_df.set_index('id', inplace=True)

for idx, name in zip(voice_id_list, name_list):
    birds_voice_df.ix[idx, 'voice'] = name
    birds_voice_df.ix[idx, 'local'] = 1

# birds_voice_df.to_json('final_set/voice.json')
birds_voice_df.to_csv('final_set/voice.csv')

birds_description_df = birds_df[['id', 'common_name', 'binomial', 'category', 'order', 'family', 
                                 'min_length', 'max_length', 'min_weight', 'max_weight', 'descriptions', 'diet', 
                                 'sources']].copy()

# birds_description_df.to_json('final_set/birds.json')
birds_description_df.set_index('id', inplace=True)
birds_description_df.to_csv('final_set/birds.csv')
